finished part 2

+++ src/main/java/Models/Shoes.java
@@ -1,11 +1,14 @@
 package Models;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class Shoes {
-   private int id;
-  private   String name;
-   private String Brand;
-  private   Float size;
-   private String sport;
+    private int id;
+    private String name;
+    private String Brand;
+    private int size;
+    private String sport;
     private int quantity;
     private double price;
 
@@ -33,11 +36,11 @@ public class Shoes {
         Brand = brand;
     }
 
-    public Float getSize() {
+    public int getSize() {
         return size;
     }
 
-    public void setSize(Float size) {
+    public void setSize(int size) {
         this.size = size;
     }
 
@@ -65,7 +68,7 @@ public class Shoes {
         this.price = price;
     }
 
-    public Shoes(int id, String name, String brand, Float size, String sport, int quantity, double price) {
+    public Shoes(int id, String name, String brand, int size, String sport, int quantity, double price) {
         this.id = id;
         this.name = name;
         Brand = brand;
@@ -75,3 +78,14 @@ public class Shoes {
         this.price = price;
     }
 }
+
+//    List<Shoes> shoesList = new ArrayList<Shoes>();
+//
+//    Shoes s1 = new Shoes(2,"Retros", "Jays", 10, "Hiking", 1, 99.00);
+//    Shoes s2 = new Shoes(4, "SB", "Nike", 8, "Running", 3, 80.00);
+//    Shoes s3 = new Shoes(1, "Yeezy", "Adidas", 9, "Flexing", 8,450.00);
+//
+//shoesList.add(s1);
+//shoesList.add(s2);
+//shoesList.add(s3);
+//}
